requirements.txt
dj-database-url==0.5.0
Django==2.2.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
requests==2.22.0
whitenoise==4.1.2
tensorflow==2.0.0b1
scikit-learn==0.21.2
numpy==1.16.3
pandas==0.24.2
cryptography==2.7
djangorestframework==3.10.1